+++ model_zoo/lbann_dnn_multi_imagenet_cv.cpp
@@ -59,8 +59,8 @@ const string g_ImageNet_LabelDir = "labels/";
 const string g_ImageNet_TrainLabelFile = "train_c0-9.txt";
 const string g_ImageNet_ValLabelFile = "val.txt";
 const string g_ImageNet_TestLabelFile = "val_c0-9.txt"; //"test.txt";
-const uint g_ImageNet_Width = 256;
-const uint g_ImageNet_Height = 256;
+const int g_ImageNet_Width = 256;
+const int g_ImageNet_Height = 256;
 
 int main(int argc, char *argv[]) {
   // El initialization (similar to MPI_Init)
@@ -209,14 +209,14 @@ int main(int argc, char *argv[]) {
       std::make_pair(execution_mode::validation, &imagenet_validation_set),
       std::make_pair(execution_mode::testing, &imagenet_testset)
     };
-    //input_layer *input_layer = new input_layer_distributed_minibatch(comm, (int) trainParams.MBSize, &imagenet_trainset, &imagenet_testset);
-    input_layer *input_layer = new input_layer_distributed_minibatch_parallel_io(data_layout::MODEL_PARALLEL, comm, parallel_io, (int) trainParams.MBSize, data_readers);
+    //input_layer *input_layer = new input_layer_distributed_minibatch(comm, trainParams.MBSize, &imagenet_trainset, &imagenet_testset);
+    input_layer *input_layer = new input_layer_distributed_minibatch_parallel_io(data_layout::MODEL_PARALLEL, comm, parallel_io, trainParams.MBSize, data_readers);
     dnn.add(input_layer);
     int NumLayers = netParams.Network.size();
     // initalize neural network (layers)
-    std::unordered_set<uint> layer_indices;
-    for (int l = 0; l < (int)NumLayers; l++) {
-      uint idx;
+    std::unordered_set<int> layer_indices;
+    for (int l = 0; l < NumLayers; l++) {
+      int idx;
       if (l < (int)NumLayers-1) {
         idx = dnn.add("FullyConnected", data_layout::MODEL_PARALLEL, netParams.Network[l],
                       trainParams.ActivationType,
@@ -231,8 +231,8 @@ int main(int argc, char *argv[]) {
       }
       layer_indices.insert(idx);
     }
-    //target_layer *target_layer = new target_layer_distributed_minibatch(comm, (int) trainParams.MBSize, &imagenet_trainset, &imagenet_testset, true);
-    target_layer *target_layer = new target_layer_distributed_minibatch_parallel_io(data_layout::MODEL_PARALLEL, comm, parallel_io, (int) trainParams.MBSize, data_readers, true);
+    //target_layer *target_layer = new target_layer_distributed_minibatch(comm, trainParams.MBSize, &imagenet_trainset, &imagenet_testset, true);
+    target_layer *target_layer = new target_layer_distributed_minibatch_parallel_io(data_layout::MODEL_PARALLEL, comm, parallel_io, trainParams.MBSize, data_readers, true);
     dnn.add(target_layer);
 
     lbann_summary summarizer(trainParams.SummaryDir, comm);
